+++ examples/meetups/H2OWorld.script.scala
@@ -1,30 +1,33 @@
+//val sc:org.apache.spark.SparkContext = null
 /**
- * Launch following commands:
- *    export MASTER="local-cluster[3,2,4096]"
- *   bin/sparkling-shell -i examples/scripts/mlconf_2015_hamSpam.script.script.scala
- *
- * When running using spark shell or using scala rest API:
- *    SQLContext is available as sqlContext
- *    SparkContext is available as sc
- */
-import _root_.hex.deeplearning.{DeepLearningModel}
+ * To start Sparkling Water please type
+
+cd path/to/sparkling/water
+export SPARK_HOME=/Users/zelleta/spark-1.2.1-bin-hadoop1
+export MASTER="local-cluster[3,2,4096]"
+
+bin/sparkling-shell --conf spark.executor.memory=3G
+*/
+
+
+// Input data
+val DATAFILE="examples/smalldata/smsData.txt"
+
+import hex.deeplearning.{DeepLearningModel, DeepLearning}
+import hex.deeplearning.DeepLearningModel.DeepLearningParameters
 import org.apache.spark.examples.h2o.DemoUtils._
 import org.apache.spark.h2o._
-import org.apache.spark.{SparkFiles, mllib}
+import org.apache.spark.mllib
 import org.apache.spark.mllib.feature.{IDFModel, IDF, HashingTF}
 import org.apache.spark.rdd.RDD
-import org.apache.spark.sql.{SQLContext, DataFrame}
 import water.Key
-import water.app.ModelMetricsSupport
 
-// Register files to SparkContext
-addFiles(sc,"examples/smalldata/smsData.txt")
 // One training message
 case class SMS(target: String, fv: mllib.linalg.Vector)
 
 // Data loader
 def load(dataFile: String): RDD[Array[String]] = {
-  sc.textFile(SparkFiles.get(dataFile)).map(l => l.split("\t")).filter(r => !r(0).isEmpty)
+  sc.textFile(dataFile).map(l => l.split("\t")).filter(r => !r(0).isEmpty)
 }
 
 // Tokenizer
@@ -63,10 +66,8 @@ def buildDLModel(train: Frame, valid: Frame,
                 (implicit h2oContext: H2OContext): DeepLearningModel = {
   import h2oContext._
   // Build a model
-  import _root_.hex.deeplearning.DeepLearning
-  import _root_.hex.deeplearning.DeepLearningParameters
   val dlParams = new DeepLearningParameters()
-  dlParams._model_id = Key.make("dlModel.hex")
+  dlParams._destination_key = Key.make("dlModel.hex").asInstanceOf[water.Key[Frame]]
   dlParams._train = train
   dlParams._valid = valid
   dlParams._response_column = 'target
@@ -85,18 +86,17 @@ def buildDLModel(train: Frame, valid: Frame,
   dlModel
 }
 
-// Create SQL support
-implicit val sqlContext = SQLContext.getOrCreate(sc)
-import sqlContext.implicits._
-
 // Start H2O services
 import org.apache.spark.h2o._
 implicit val h2oContext = new H2OContext(sc).start()
 import h2oContext._
-
+// Initialize SQL context
+import org.apache.spark.sql._
+implicit val sqlContext = new SQLContext(sc)
+import sqlContext._
 
 // Data load
-val data = load("smsData.txt")
+val data = load(DATAFILE)
 // Extract response spam or ham
 val hamSpam = data.map( r => r(0))
 val message = data.map( r => r(1))
@@ -107,11 +107,9 @@ val tokens = tokenize(message)
 var (hashingTF, idfModel, tfidf) = buildIDFModel(tokens)
 
 // Merge response with extracted vectors
-val resultRDD: DataFrame = hamSpam.zip(tfidf).map(v => SMS(v._1, v._2)).toDF
+val resultRDD: SchemaRDD = hamSpam.zip(tfidf).map(v => SMS(v._1, v._2))
 
-val table:H2OFrame = resultRDD
-// Transform target column into
-table.replace(table.find("target"), VecUtils.stringToCategorical(table.vec("target"))).remove()
+val table:DataFrame = resultRDD
 
 // Split table
 val keys = Array[String]("train.hex", "valid.hex")
@@ -123,33 +121,38 @@ table.delete()
 // Build a model
 val dlModel = buildDLModel(train, valid)
 
-/*
- * The following code is appended life during presentation.
- */
 // Collect model metrics and evaluate model quality
-val trainMetrics = ModelMetricsSupport.binomialMM(dlModel, train)
-val validMetrics = ModelMetricsSupport.binomialMM(dlModel, valid)
-println(trainMetrics.auc._auc)
-println(validMetrics.auc._auc)
+val trainMetrics = binomialMM(dlModel, train)
+val validMetrics = binomialMM(dlModel, valid)
+println(trainMetrics.auc.AUC)
+println(validMetrics.auc.AUC)
 
 // Spam detector
 def isSpam(msg: String,
            dlModel: DeepLearningModel,
            hashingTF: HashingTF,
            idfModel: IDFModel,
-           h2oContext: H2OContext,
            hamThreshold: Double = 0.5):Boolean = {
   val msgRdd = sc.parallelize(Seq(msg))
-  val msgVector: DataFrame = idfModel.transform(
+  val msgVector: SchemaRDD = idfModel.transform(
                               hashingTF.transform (
-                                tokenize (msgRdd))).map(v => SMS("?", v)).toDF
-  val msgTable: H2OFrame = h2oContext.asH2OFrame(msgVector)
+                                tokenize (msgRdd))).map(v => SMS("?", v))
+  val msgTable: DataFrame = msgVector
   msgTable.remove(0) // remove first column
   val prediction = dlModel.score(msgTable)
   //println(prediction)
   prediction.vecs()(1).at(0) < hamThreshold
 }
 
-println(isSpam("Michal, h2oworld party tonight in MV?", dlModel, hashingTF, idfModel, h2oContext))
-println(isSpam("We tried to contact you re your reply to our offer of a Video Handset? 750 anytime any networks mins? UNLIMITED TEXT?", dlModel, hashingTF, idfModel, h2oContext))
+println(isSpam("Michal, beer tonight in MV?", dlModel, hashingTF, idfModel))
+println(isSpam("We tried to contact you re your reply to our offer of a Video Handset? 750 anytime any networks mins? UNLIMITED TEXT?", dlModel, hashingTF, idfModel))
 
+//
+// This is R code which should be put into R session:
+//
+// library(h2o)
+// h2o.init()
+// dlModel = h2o.getModel("dlModel.hex")
+// validation.df = as.data.frame(t(runif(n = 1024, min = 0, max = 4)))
+// names(validation.df) <- paste0("a",0:1023)
+// predict(dlModel, validation.df)