+++ py/clean.sh
@@ -1,5 +1,0 @@
-#!/usr/bin/env bash
-rm -rf sparkling_water.egg-info/
-rm -rf build/
-rm -rf dist/
-rm -rf data_files/